Finished step 1

+++ compile.sh
@@ -1,2 +1,4 @@
-rm ./log-client.exe
+if [ -f ./log-client.exe ]; then
+	rm ./log-client.exe
+fi
 g++ ./src/main.cpp -o log-client.exe
